default force rerun to false
As we're talking about it several times before I propose this solution.
For now even when we add an extra info to the logger all steps
will be re-run. As we provide the rerun flag I thik it'd be beneficial
for everyone of we default the rerun to false and just infor the user
about the need of the rerun when the model code was changed.

If you feel differently feel free to open the discussion below

#155